+++ cpp/src/scheduler/resource/Node.cpp
@@ -17,27 +17,6 @@ Node::Node() {
     id_ = counter++;
 }
 
-void Node::DelNeighbour(const NeighbourNodePtr &neighbour_ptr) {
-    std::lock_guard<std::mutex> lk(mutex_);
-    if (auto s = neighbour_ptr.lock()) {
-        auto search = neighbours_.find(s->id_);
-        if (search != neighbours_.end()) {
-            neighbours_.erase(search);
-        }
-    }
-}
-
-bool Node::IsNeighbour(const NeighbourNodePtr &neighbour_ptr) {
-    std::lock_guard<std::mutex> lk(mutex_);
-    if (auto s = neighbour_ptr.lock()) {
-        auto search = neighbours_.find(s->id_);
-        if (search != neighbours_.end()) {
-            return true;
-        }
-    }
-    return false;
-}
-
 std::vector<Neighbour> Node::GetNeighbours() {
     std::lock_guard<std::mutex> lk(mutex_);
     std::vector<Neighbour> ret;
@@ -48,8 +27,13 @@ std::vector<Neighbour> Node::GetNeighbours() {
 }
 
 std::string Node::Dump() {
-    // TODO(linxj): what's that?
-    return std::__cxx11::string();
+    std::stringstream ss;
+    ss << "<Node, id=" << std::to_string(id_) << ">::neighbours:" << std::endl;
+    for (auto &neighbour : neighbours_) {
+        ss << "\t<Neighbour, id=" << std::to_string(neighbour.first);
+        ss << ", connection: " << neighbour.second.connection.Dump() << ">" << std::endl;
+    }
+    return ss.str();
 }
 
 void Node::AddNeighbour(const NeighbourNodePtr &neighbour_node, Connection &connection) {